Fix movement when resizing page
Separating the header and hero worked! now it moves as intended and
adding the max-width to the header worked as I wanted.

note: its better to separate every section, even if the flex:column
worked for displaying the header and hero it introduced limitations to
the movement of the objects.